execution-core: Add rusk-abi types and functionality
