+++ execution-core/Cargo.toml
@@ -10,6 +10,7 @@ dusk-poseidon = "0.39"
 bls12_381-bls = { version = "0.4", default-features = false, features = ["rkyv-impl"] }
 jubjub-schnorr = { version = "0.4", default-features = false, features = ["rkyv-impl"] }
 phoenix-core = { version = "0.30.0-rc", default-features = false, features = ["rkyv-impl", "alloc"] }
+piecrust-uplink = { version = "0.16" }
 dusk-bytes = "0.1"
 rkyv = { version = "0.7", default-features = false,  features = ["size_32"] }
 bytecheck = { version = "0.6", default-features = false }
