+++ execution-core/src/lib.rs
@@ -11,13 +11,16 @@
 #![deny(rustdoc::broken_intra_doc_links)]
 #![deny(clippy::pedantic)]
 #![allow(clippy::module_name_repetitions)]
+#![feature(const_fn_floating_point_arithmetic)]
 
 extern crate alloc;
 
-pub mod bytecode;
-pub mod reader;
+pub mod license;
 pub mod stake;
 pub mod transfer;
+
+mod dusk;
+pub use dusk::{dusk, from_dusk, Dusk, LUX};
 
 // elliptic curve types
 pub use dusk_bls12_381::BlsScalar;
@@ -26,28 +29,85 @@ pub use dusk_jubjub::{
     GENERATOR_NUMS_EXTENDED,
 };
 
-// signature types
-pub use bls12_381_bls::{
-    Error as BlsSigError, MultisigPublicKey as BlsMultisigPublicKey,
-    MultisigSignature as BlsMultisigSignature, PublicKey as BlsPublicKey,
-    SecretKey as BlsSecretKey, Signature as BlsSignature,
+/// Signatures used in the Dusk protocol.
+pub mod signatures {
+    /// Types for the bls-signature scheme.
+    pub mod bls {
+        pub use bls12_381_bls::{
+            Error, MultisigPublicKey, MultisigSignature, PublicKey, SecretKey,
+            Signature,
+        };
+    }
+
+    /// Types for the schnorr-signature scheme.
+    pub mod schnorr {
+        pub use jubjub_schnorr::{
+            PublicKey, SecretKey, Signature, SignatureDouble,
+        };
+    }
+}
+
+pub use piecrust_uplink::{
+    ContractError, ContractId, Event, StandardBufSerializer, ARGBUF_LEN,
+    CONTRACT_ID_BYTES,
 };
 
-pub use jubjub_schnorr::{
-    PublicKey as SchnorrPublicKey, SecretKey as SchnorrSecretKey,
-    Signature as SchnorrSignature, SignatureDouble as SchnorrSignatureDouble,
-};
+#[inline]
+const fn reserved(b: u8) -> ContractId {
+    let mut bytes = [0u8; CONTRACT_ID_BYTES];
+    bytes[0] = b;
+    ContractId::from_bytes(bytes)
+}
 
-/// Secret key associated with a note.
-pub type NoteSecretKey = SchnorrSecretKey;
-/// Public key associated with a note.
-pub type NotePublicKey = SchnorrPublicKey;
-/// Signature to prove ownership of the note
-pub type NoteSignature = SchnorrSignature;
+use alloc::string::String;
+use alloc::vec::Vec;
 
-// phoenix types
-pub use phoenix_core::{
-    value_commitment, Error as PhoenixError, Note, PublicKey, SecretKey,
-    Sender, StealthAddress, TxSkeleton, ViewKey, NOTE_VAL_ENC_SIZE,
-    OUTPUT_NOTES,
-};
+use dusk_bytes::{DeserializableSlice, Error as BytesError};
+
+/// Reads vector from a buffer.
+/// Resets buffer to a position after the bytes read.
+///
+/// # Errors
+/// When length or data could not be read.
+fn read_vec(buf: &mut &[u8]) -> Result<Vec<u8>, BytesError> {
+    let len = usize::try_from(u64::from_reader(buf)?)
+        .map_err(|_| BytesError::InvalidData)?;
+    if buf.len() < len {
+        return Err(BytesError::InvalidData);
+    }
+    let bytes = buf[..len].into();
+    *buf = &buf[len..];
+    Ok(bytes)
+}
+
+/// Reads string from a buffer.
+/// Resets buffer to a position after the bytes read.
+///
+/// # Errors
+/// When length or data could not be read.
+fn read_str(buf: &mut &[u8]) -> Result<String, BytesError> {
+    let len = usize::try_from(u64::from_reader(buf)?)
+        .map_err(|_| BytesError::InvalidData)?;
+    if buf.len() < len {
+        return Err(BytesError::InvalidData);
+    }
+    let str = String::from_utf8(buf[..len].into())
+        .map_err(|_| BytesError::InvalidData)?;
+    *buf = &buf[len..];
+    Ok(str)
+}
+
+/// Reads array from a buffer.
+/// Resets buffer to a position after the bytes read.
+///
+/// # Errors
+/// When length or data could not be read.
+fn read_arr<const N: usize>(buf: &mut &[u8]) -> Result<[u8; N], BytesError> {
+    if buf.len() < N {
+        return Err(BytesError::InvalidData);
+    }
+    let mut a = [0u8; N];
+    a.copy_from_slice(&buf[..N]);
+    *buf = &buf[N..];
+    Ok(a)
+}
